Add service client configuration and setup
This commit introduces service client setup with the addition of `ServiceClientsExtension.cs` file in the `Infrastructure/Extensions` directory. The file contains `AddServiceClients` & `AddClientSettings` methods, which sets up the HTTP client for consuming another microservice and configures it with the settings from `appsettings.json`. The `Configuration` edge is added for these clients in the `Infrastructure/DependencyInjection.cs`. Moreover, app settings are updated to include configurations for another microservice endpoint. This is critical to ensure scalability and loose coupling between services.