Remove unnecessary files and clean up repository
Deleted several files including auto-generated, cache, and configuration files which were not necessary in the repository. They are potentially transient files that should be generated by IDEs or build systems and therefore should not be tracked in the version control system. This cleanup helps to decrease repository size and also avoid potential conflicting changes that may occur when multiple people are working on the same project concurrently. Moreover, the removal of these non-essential files contributes in making the repository more readable and maintainable.