Refactor DTO Namespace and Improve Error Handling
This commit centralizes DTO (Data Transfer Objects) into a common namespace which simplifies the references in other files and makes the architecture more coherent. Previously, DTOs were scattered in different places with unnecessary complexity.

Also, an enhancement was made for better visibility on error handling within the HttpClient by adding Console.WriteLine in the catch block. It helps to diagnose issues faster when an exception is thrown.

Note: Some other files also had to be updated to match the new DTO paths.