Refactor code and update connection configuration
This commit removed the redundant service 'AddMvc()' in 'Program.cs' as it was unnecessary. Changed the way endpoints are being used for better efficiency. Also, updated 'appsettings.json' to use local host for development reasons. Modified the RedisCache configuration in 'RedisExtension.cs' and 'RedisCacheConfigurations.cs' to streamline usage. In 'CreateUserV1CommandHandler.cs', made an update to save the 'request.Name' instead of hard-coded string. Lastly, I added an 'Id' field in 'TestResponse' for tracking purposes.